+++ ball.h
@@ -16,8 +16,8 @@ public:
     int get_Amount() const { return amount; }
     std::vector<std::string> get_All_Name() const { return name; }
     std::vector<Cartesian> get_All_Pos() const { return position; }
-    int get_Ball_Index(std::string str) const;
-    std::string get_Ball_Name(int index) const { return name[index]; }
+    size_t get_Ball_Index(std::string str) const;
+    std::string get_Ball_Name(size_t index) const { return name[index]; }
 
     void set_Amount(int num) { amount = num; }
     void set_All_Name(std::vector<std::string> names) { name = names; }
